noteDelete validated!

+++ src/main/java/lk/ijse/gdse/aad68/notetaker/service/NoteServiceIMPL.java
@@ -35,8 +35,13 @@ public  class NoteServiceIMPL implements NoteService {
     }
 
     @Override
-    public void deleteNote(String noteId) {
-        noteDao.deleteById(noteId);
+    public boolean deleteNote(String noteId) {
+        if (noteDao.existsById(noteId)){
+            noteDao.deleteById(noteId);
+            return true;
+        }else {
+            return false;
+        }
 
     }
